Redo entity system
Entities are now big monolithic world objects. Any new entity feature
must be added to the Entity class, and these features can be enabled or
disabled. This is similar to a component-entity system, but the
difference is that components are all contained in one Entity class for
better garbage collection performance.